floatfield

+++ Linker/models.py
@@ -5,8 +5,8 @@ from django.contrib.auth.models import User
 # Create your models here.
 class SensorData(models.Model):
     Location = models.CharField(max_length=100)
-    Voltage_value = models.IntegerField() 
-    Current_value = models.IntegerField() 
+    Voltage_value = models.FloatField() 
+    Current_value = models.FloatField() 
     date_posted = models.DateTimeField(default=timezone.now)
     author = models.ForeignKey(User, on_delete=models.CASCADE)
